requirements.txt
Scrapy~=2.4.0
itemadapter~=0.1.1
PyPDF2~=1.26.0
pandas~=1.1.3
tabula-py~=2.2.0